separate executions for mojarra and myfaces
