+++ jsf-spring-boot-autoconfigure/src/test/java/org/joinfaces/myfaces/MyfacesJsfClassFactoryConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2016 the original author or authors.
+ * Copyright 2016-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package org.joinfaces;
+package org.joinfaces.myfaces;
 
 import javax.servlet.ServletContainerInitializer;
 
 import org.apache.myfaces.ee6.MyFacesContainerInitializer;
+import org.joinfaces.JsfClassFactoryConfiguration;
 
 public class MyfacesJsfClassFactoryConfiguration implements JsfClassFactoryConfiguration {
